pyproject.toml
# pyproject.toml (only the relevant parts)

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[project]
name = "spotifyai"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "fastmcp>=2.12.3",
  "spotipy>=2.25.1",
  "mcp>=1.14.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "mypy>=1.11.2",
  "types-requests>=2.32.0.20240914",
]

[project.scripts]
spotify-mcp = "mcp_server.spotify_server:main"

# ---------- Ruff ----------
[tool.ruff]
target-version = "py313"    # match your .tool-versions (3.13.7)
line-length = 100
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# ---------- mypy ----------
[tool.mypy]
python_version = "3.13"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true  # quiet for libs without stubs
pretty = true

[[tool.mypy.overrides]]
module = ["mcp_server.*", "main"]
strict = false
disallow_untyped_defs = true
check_untyped_defs = true